07-streaming/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
langchain-anthropic==0.3.0
langgraph==0.2.55
langgraph-checkpoint-sqlite==2.0.1
aiosqlite==0.20.0
